fixed authentication

+++ controllers/gif.js
@@ -29,7 +29,8 @@ exports.getOneGif = (req, res) => {
 
 exports.postGif = (req, res) => {
   const file = req.files[0].path;
-  const { title, userId } = req.body;
+  const { title, userId, category } = req.body;
+  // console.log(req.body.userId);
   // Upload file to Cloudinary
   cloudinary.upload(file)
     .then((image) => {
@@ -37,8 +38,8 @@ exports.postGif = (req, res) => {
         image.url,
         image.public_id
       ];
-      query(`INSERT INTO feeds (Title, Content, UserID, Type )
-      VALUES ($1, $2, $3, 'gif') RETURNING *;`, [title, imageDetails, userId])
+      query(`INSERT INTO feeds (Title, Content, UserID, Type, Category, IsFlagged )
+      VALUES ($1, $2, $3, 'gif', $4, false ) RETURNING *;`, [title, imageDetails, userId, category])
         .then((result) => res.status(201).json({
           status: 'success',
           data: {
@@ -62,9 +63,9 @@ exports.postGif = (req, res) => {
 
 exports.deleteGif = (req, res) => {
   const gifId = req.params.id;
-  const publicId = req.body.cloudId;
-  cloudinary.delete(publicId)
-    .then(() => query("DELETE FROM feeds WHERE (id=$1 AND type='gif')", [gifId])
+  const { userId, cloudId } = req.body;
+  cloudinary.delete(cloudId)
+    .then(() => query("DELETE FROM feeds WHERE (id=$1 AND UserID=$2 AND type='gif')", [gifId, userId])
       .then(() => res.status(200).json({ message: 'GIF was successfully deleted' })))
     .catch((err) => res.status(400).json({ error: err }));
 };
